cleanup and fixing tests
